refactor(mcp): move connector catalog to interactive tools
